Move the controller models outside of the form package.
